Added cross-platform include for SDL, SDL_include.h. Have been trying to get it to compile on Windows, it does now, but apparently the heap is corrupted upon startup.... after 2 hours, I give up.... For now.

+++ src/game/GameEngine.cpp
@@ -36,7 +36,7 @@ void GameEngine::InitData()
 
 }
 
-void GameEngine::tic(unsigned int td) {
+void GameEngine::tic(uint64_t td) {
    // Update functions go here
    m_player->tic(td);
    m_world->update(m_player->getPosition());
